ASU COK:

+++ db/migrate/20190220234848_create_table_order.rb
@@ -7,10 +7,10 @@ class CreateTableOrder < ActiveRecord::Migration[5.2]
 		  	t.references :supplier, foreign_key: true, null: false
 		  	t.references :store, foreign_key: true, null: false
 		  	t.integer :total_items, null: false
-		  	t.integer :total, null: false
+		  	t.float :total, null: false
 		  	t.datetime :date_paid_off
 		  	t.boolean :editable, null: false, default: true
-		  	t.bigint :old_total, null: false, default: 0
+		  	t.float :old_total, null: false, default: 0
 		  	t.datetime :date_change
 		  	t.references :user, foreign_key: true, null: false
 		  	t.bigint :received_by
